Extract endpoints for secrets and assets to separate files
