+++ public/scripts/secrets.js
@@ -73,7 +73,7 @@ export let secret_state = {};
 
 export async function writeSecret(key, value) {
     try {
-        const response = await fetch('/writesecret', {
+        const response = await fetch('/api/secrets/write', {
             method: 'POST',
             headers: getRequestHeaders(),
             body: JSON.stringify({ key, value }),
@@ -94,7 +94,7 @@ export async function writeSecret(key, value) {
 
 export async function readSecretState() {
     try {
-        const response = await fetch('/readsecretstate', {
+        const response = await fetch('/api/secrets/read', {
             method: 'POST',
             headers: getRequestHeaders(),
         });
